Ready for release?

+++ holdem/appsrc/testDriver.cpp
@@ -555,9 +555,9 @@ if( bLoadGame )
 
         const uint32 NUM_OPPONENTS = 8;
         const uint32 rand765 = 1 + ((blindIncrFreq + tokenRandomizer)^(blindIncrFreq*tokenRandomizer));
-        const uint32 rand8432 = 1 + (labs(blindIncrFreq - tokenRandomizer)^(blindIncrFreq*tokenRandomizer));
+        const uint64_t rand8432 = 1 + (labs(blindIncrFreq - tokenRandomizer)^(blindIncrFreq*tokenRandomizer));
         const uint32 rand8 = rand8432%8;
-        const uint32 rand432 = rand8432/8;
+        const uint32 rand432 = static_cast<uint32>(rand8432/8);
         const uint32 randSeed = rand8 + (rand765%(7*6*5))*8 + (rand432%(4*3*2))*(8*7*6*5);
         uint8 i;
         int8 * opponentorder;
